+++ src/api-rest/controllers/action.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const services = require('../services');
+const services = require('../../services');
 const mongoose = require('mongoose');
 
 /**
@@ -13,7 +13,7 @@ async function holdProduct(req, res, next) {
     if (amount <= 0) {
       next('Amount has to be greater than 0');
     } else {
-      const { accountId, productId } = req.params;
+      const { accountId, productId } = req.query;
       if (amount && mongoose.Types.ObjectId.isValid(accountId) && mongoose.Types.ObjectId.isValid(productId)) {
         await services.queries.holdProduct({ amount, accountId, productId });
         res.status(200).json({ 'request-id': req.header('X-REQUEST-ID') });
@@ -33,7 +33,7 @@ async function holdProduct(req, res, next) {
 async function updateCartAmount(req, res, next) {
   try {
     const { amount } = req.body;
-    const { accountId, productId } = req.params;
+    const { accountId, productId } = req.query;
     if (mongoose.Types.ObjectId.isValid(accountId) && mongoose.Types.ObjectId.isValid(productId)) {
       await services.queries.updateCartAmount({ amount, accountId, productId });
       res.status(200).json({ 'request-id': req.header('X-REQUEST-ID') });
@@ -51,7 +51,7 @@ async function updateCartAmount(req, res, next) {
  */
 async function moveCart(req, res, next) {
   try {
-    const { accountId, productId } = req.params;
+    const { accountId, productId } = req.query;
     if (mongoose.Types.ObjectId.isValid(accountId) && mongoose.Types.ObjectId.isValid(productId)) {
       await services.queries.moveCart({ accountId, productId });
       res.status(200).json({ 'request-id': req.header('X-REQUEST-ID') });