+++ src/api-rest/controllers/index.js
@@ -2,7 +2,7 @@
 
 const express = require('express');
 const basicAuth = require('express-basic-auth');
-const { server } = require('./../config');
+const { server } = require('./../../config');
 const router = express.Router();
 
 const { addAccount, removeAccount, getAccountHolds } = require('./account');
@@ -22,9 +22,9 @@ function controllers() {
   router.patch('/product/:productId', basicAuth(basicAuthObject), updateProductStock);
   router.get('/products', listProducts);
 
-  router.post('/action/:accountId/product/:productId', holdProduct);
-  router.patch('/action/:accountId/product/:productId', updateCartAmount);
-  router.delete('/action/:accountId/product/:productId', moveCart);
+  router.post('/action/hold', holdProduct);
+  router.patch('/action/hold', updateCartAmount);
+  router.delete('/action/hold', moveCart);
 
   return router;
 }