+++ src/config/server.js
@@ -3,11 +3,12 @@
 const Joi = require('@hapi/joi');
 
 const envVarsSchema = Joi.object({
+  API_TYPE: Joi.string().valid('rest', 'graphql').default('graphql'),
   API_PROTOCOL: Joi.string().valid('http', 'https').default('http'),
   API_PORT: Joi.number().default(8080),
   API_ADMIN_LOGIN: Joi.string().default('admin'),
   API_ADMIN_PASSWORD: Joi.string().default('admin'),
-  API_APIKEY_VALUE: Joi.string().default('apikey ABCD'),
+  API_APIKEY_VALUE: Joi.string(),
   API_SSL_KEY: Joi.string(),
   API_SSL_CERT: Joi.string()
 })
@@ -20,6 +21,7 @@ if (error) {
 }
 
 module.exports = {
+  apiType: envVars.API_TYPE,
   protocol: envVars.API_PROTOCOL,
   port: envVars.API_PORT,
   basicAuth: {